Remove the sitemaps
The sitemap applications that power these sitemaps are about to be disabled to avoid a forced switch off for using a deprecated API in Appengine (Files and Blob storage).

There is a discussion going on about the future of the sitemaps and if they are to be revived in future this change can just be reverted.